+++ src/ts/Ana/Ana.ts
@@ -13,6 +13,7 @@ import { rSubheading } from '../Components/Atoms/Subheading/Subheading'
 import { rParagraph } from '../Components/Atoms/Paragraph/Paragraph'
 import { rLabel } from '../Components/Atoms/Label/Label'
 import { rLink } from '../Components/Atoms/Link/Link'
+import { rCheckbox } from '../Components/Molecules/Checkbox/Checkbox'
 
 declare global {
   interface HTMLElement {
@@ -110,13 +111,15 @@ export class Ana implements iAna {
       Paragraph: rParagraph(elements, this.configuration),
       Label: rLabel(elements, this.configuration),
       Link: rLink(elements, this.configuration),
-      //   Display: rDisplay(elements, this.configuration),
+    }
+    let molecules = {
+      Checkbox: rCheckbox(elements, this.configuration),
     }
     // let organisms = {
     //   Testpage: rTestpage(elements, this.configuration)
     // }
     // return { ...elements, ...atoms, ...organisms }
-    return { ...elements, ...atoms}
+    return { ...elements, ...atoms, ...molecules}
   }
 
   /**